refactor: decompose components and fix prisma 7 engine init
- extracted mapbox map logic into modular utilities under lib/map/
- created useSearchSuggestions hook to separate logic from UI
- resolved error with prisma integration ValidationError
- updated schema.prisma to use engineType = client
- standardized all Prisma client signleton in lib/prisma.ts
- cleaned up API routes and improved type safety in auth and profile
  flows
- removed unecessary comments and console logs
