+++ next.config.ts
@@ -14,6 +14,7 @@ const nextConfig: NextConfig = {
       },
     ],
   },
+  serverExternalPackages: ["@prisma/client"],
 };
 
 export default nextConfig;
