+++ package.json
@@ -7,7 +7,8 @@
     "build": "next build --turbopack",
     "start": "next start",
     "lint": "eslint",
-    "db:generate": "dotenv -e .env.local -- prisma generate",
+    "postinstall": "prisma generate",
+    "db:generate": "prisma generate",
     "db:push": "dotenv -e .env.local -- prisma db push",
     "db:migrate": "dotenv -e .env.local -- prisma migrate dev",
     "db:studio": "dotenv -e .env.local -- prisma studio",
@@ -15,8 +16,8 @@
   },
   "dependencies": {
     "@mapbox/search-js-react": "^1.5.0",
-    "@prisma/client": "^6.17.1",
-    "@prisma/extension-accelerate": "^2.0.2",
+    "@prisma/adapter-pg": "^7.6.0",
+    "@prisma/client": "^7.6.0",
     "@radix-ui/react-collapsible": "^1.1.12",
     "@radix-ui/react-slider": "^1.3.6",
     "@radix-ui/react-slot": "^1.2.3",
@@ -24,6 +25,7 @@
     "@supabase/supabase-js": "^2.100.1",
     "@tailwindcss/vite": "^4.1.14",
     "@turf/turf": "^7.2.0",
+    "@types/pg": "^8.20.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dotenv": "^17.3.1",
@@ -35,6 +37,7 @@
     "lucide-react": "^0.544.0",
     "mapbox-gl": "^3.15.0",
     "next": "^15.5.12",
+    "pg": "^8.20.0",
     "react": "19.1.0",
     "react-circular-progressbar": "^2.2.0",
     "react-dom": "19.1.0",
