+++ src/lib/auth/registrant.ts
@@ -1,15 +1,13 @@
-import { PrismaClient } from "@/prisma_app/generated/prisma";
+import { prisma } from "@/lib/prisma";
 
-const prisma = new PrismaClient();
 
 export function normalizeEmail(email: string) {
   return email.trim().toLowerCase();
 }
 
-/** True if this email or Supabase user id is already registered in Prisma (User or Profile). */
 export async function registrantExists(
   email: string | undefined | null,
-  supabaseUserId: string
+  supabaseUserId: string,
 ): Promise<boolean> {
   if (!email) {
     const byId = await prisma.profile.findUnique({
@@ -35,8 +33,9 @@ export async function registrantExists(
   return !!(user || profileByEmail || profileById);
 }
 
-/** True if email is taken in Prisma (blocks new email/password sign-up). */
-export async function isEmailRegisteredInPrisma(email: string): Promise<boolean> {
+export async function isEmailRegisteredInPrisma(
+  email: string,
+): Promise<boolean> {
   const norm = normalizeEmail(email);
   const [user, profile] = await Promise.all([
     prisma.user.findFirst({
@@ -49,13 +48,9 @@ export async function isEmailRegisteredInPrisma(email: string): Promise<boolean>
   return !!(user || profile);
 }
 
-/**
- * Creates or updates a minimal Profile after email sign-up, verification, or OAuth.
- * Email may be null for some providers (e.g. Facebook) until the user adds it in onboarding.
- */
 export async function bootstrapProfileStub(
   supabaseUserId: string,
-  email: string | null | undefined
+  email: string | null | undefined,
 ) {
   const norm = email ? normalizeEmail(email) : null;
   await prisma.profile.upsert({
